pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cip-health"
version = "0.1.0"
description = "Consumer Intelligence Protocol — Personal Health domain MCP server"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "anthropic>=0.40",
    "openai>=1.50",
    "cryptography>=43.0",
    "structlog>=23.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]
# NOTE: mantic-thinking is no longer a Python dependency. CIP Health calls
# cip-mantic-core as an MCP service (MCP-to-MCP) via the ManticMCPClient.
# See: src/cip/core/mantic/client.py

[tool.hatch.build.targets.wheel]
packages = ["src/cip"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
